feat: listening actor
