+++ src/services/client/mod.rs
@@ -1,5 +1,4 @@
 pub mod arguments;
+pub mod manager;
 pub mod request;
 pub mod stream;
-pub mod manager;
-
