+++ src/services/cluster/inbound/mod.rs
@@ -1,3 +1,3 @@
-pub mod stream;
-pub mod request;
 pub mod arguments;
+pub mod request;
+pub mod stream;
