+++ src/services/cluster/inbound/stream.rs
@@ -1,5 +1,6 @@
 use crate::make_smart_pointer;
-use crate::services::cluster::actor::PeerAddr;
+
+use crate::services::cluster::actors::actor::PeerAddr;
 use crate::services::cluster::inbound::request::HandShakeRequest;
 use crate::services::cluster::inbound::request::HandShakeRequestEnum;
 
@@ -55,7 +56,7 @@ impl InboundStream {
         self.write(QueryIO::SimpleString(
             "FULLRESYNC 8371b4fb1155b71f4a04d3e1bc3e18c4a990aeeb 0".to_string(),
         ))
-            .await?;
+        .await?;
 
         Ok((repl_id, offset))
     }
